Refactors role-based access control
Replaces the `IsAdmin` middleware with a more flexible `CheckRole` middleware.
This change allows for easier implementation of different role-based permissions.

Updates the admin route group to use the new middleware with the 'admin' role.
Redirects unauthorized users to the login page instead of displaying an error.

Also, filters the galleries to only include active items.